requirements.txt
airr==1.2.1
alabaster==0.7.12
alembic==1.3.2
Babel==2.8.0
bcrypt==3.1.7
biopython==1.76
blinker==1.4
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
Click==7.0
colorama==0.4.3
commonmark==0.9.1
docutils==0.15.2
dominate==2.4.0
Flask==1.1.1
Flask-Admin==1.5.4
Flask-Babel==0.12.2
Flask-BabelEx==0.9.3
Flask-Bootstrap==3.3.7.1
Flask-DebugToolbar==0.10.1
Flask-Limiter==1.1.0
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.5.2
Flask-OAuth==0.12
Flask-Principal==0.4.0
Flask-Security==3.0.0
Flask-SQLAlchemy==2.4.1
Flask-Table==0.5.0
Flask-WTF==0.14.2
future==0.18.2
html5lib==1.0.1
httplib2==0.15.0
idna==2.8
imagesize==1.2.0
itsdangerous==1.1.0
Jinja2==2.10.3
limits==1.4.1
Mako==1.1.0
MarkupSafe==1.1.1
mysql-connector-python==8.0.18
mysqlclient==1.4.6
numpy==1.18.0
oauth2==1.9.0.post1
packaging==19.2
pandas==0.25.3
passlib==1.7.2
pip-tools==4.3.0
protobuf==3.11.2
pycparser==2.19
Pygments==2.5.2
pyparsing==2.4.6
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2019.3
PyYAML==5.2
recommonmark==0.6.0
requests==2.22.0
selenium==3.141.0
six==1.13.0
snowballstemmer==2.0.0
speaklater==1.3
Sphinx==2.3.1
sphinx-bootstrap-theme==0.7.1
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-autoprogram==0.1.5
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
sphinxcontrib-websupport==1.1.2
SQLAlchemy==1.3.12
textile==3.0.4
urllib3==1.25.7
visitor==0.1.3
webencodings==0.5.1
Werkzeug==0.16.0
wincertstore==0.2
WTForms==2.2.1
xmltodict==0.12.0
yamlordereddictloader==0.4.0
